requirement.txt
alembic==1.0.0
Babel==2.6.0
blinker==1.4
click==6.7
dominate==2.3.1
Flask==1.0.2
Flask-Admin==1.5.2
Flask-BabelEx==0.9.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
ForgeryPy==0.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.13
Pillow==5.2.0
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.11
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
